requirements.txt
# requirements.txt
pyTelegramBotAPI>=4.12.0
google-cloud-storage>=2.7.0
google-cloud-speech>=2.16.0
python-dotenv>=0.20.0